Move sketch sweeping code to dedicated module
